+++ common/src/main/java/com/faboslav/friendsandfoes/mixin/RaidMixin.java
@@ -25,9 +25,11 @@ public class RaidMixin
 		if (
 			(
 				raider instanceof IllusionerEntity
+				&& FriendsAndFoes.getConfig().enableIllusioner == false
 				&& FriendsAndFoes.getConfig().enableIllusionerInRaids == false
 			) || (
 				raider instanceof IceologerEntity
+				&& FriendsAndFoes.getConfig().enableIceologer == false
 				&& FriendsAndFoes.getConfig().enableIceologerInRaids == false
 			)
 		) {