+++ fabric/src/main/java/com/faboslav/friendsandfoes/fabric/EarlyRiserFabric.java
@@ -1,7 +1,9 @@
 package com.faboslav.friendsandfoes.fabric;
 
 import com.chocohead.mm.api.ClassTinkerers;
+import com.faboslav.friendsandfoes.FriendsAndFoes;
 import com.faboslav.friendsandfoes.init.ModEntity;
+import com.faboslav.friendsandfoes.util.ExpandedEnumValues;
 import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.MappingResolver;
 
@@ -16,8 +18,15 @@ public final class EarlyRiserFabric implements Runnable
 
 		ClassTinkerers
 			.enumBuilder(Raid, EntityType, int[].class)
-			.addEnum("ILLUSIONER", () -> new Object[]{net.minecraft.entity.EntityType.ILLUSIONER, new int[]{0, 0, 0, 0, 1, 0, 1, 1}})
-			.addEnum("ICEOLOGER", () -> new Object[]{ModEntity.ICEOLOGER.get(), new int[]{0, 0, 0, 0, 1, 1, 0, 1}})
+			.addEnum(ExpandedEnumValues.ILLUSIONER, () -> new Object[]{net.minecraft.entity.EntityType.ILLUSIONER, new int[]{0, 0, 0, 0, 1, 0, 1, 1}})
+			.addEnum(ExpandedEnumValues.ICEOLOGER, () -> new Object[]{ModEntity.ICEOLOGER.get(), new int[]{0, 0, 0, 0, 1, 1, 0, 1}})
+			.build();
+
+		String SpawnGroup = remapper.mapClassName("intermediary", "net.minecraft.class_1311");
+
+		ClassTinkerers
+			.enumBuilder(SpawnGroup, String.class, int.class, boolean.class, boolean.class, int.class)
+			.addEnum(ExpandedEnumValues.GLARES, () -> new Object[]{FriendsAndFoes.makeStringID("glares"), 5, true, false, 128})
 			.build();
 	}
 }