+++ forge/src/main/java/com/faboslav/friendsandfoes/forge/FriendsAndFoesForgeClient.java
@@ -1,8 +1,13 @@
 package com.faboslav.friendsandfoes.forge;
 
+import com.faboslav.friendsandfoes.FriendsAndFoes;
 import com.faboslav.friendsandfoes.FriendsAndFoesClient;
+import com.faboslav.friendsandfoes.config.ConfigScreenBuilder;
 import dev.architectury.platform.Platform;
 import dev.architectury.utils.Env;
+import net.minecraftforge.client.ConfigGuiHandler;
+import net.minecraftforge.fml.ModList;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
 public final class FriendsAndFoesForgeClient
@@ -22,6 +27,14 @@ public final class FriendsAndFoesForgeClient
 			}
 
 			FriendsAndFoesClient.initCustomRegisters();
+
+			if(ModList.get().isLoaded("cloth_config")) {
+				ModLoadingContext.get().registerExtensionPoint(ConfigGuiHandler.ConfigGuiFactory.class, () ->
+					new ConfigGuiHandler.ConfigGuiFactory(
+						(mc, screen) -> ConfigScreenBuilder.createConfigScreen(FriendsAndFoes.getConfig(), screen)
+					)
+				);
+			}
 		});
 	}
 }